perf: wire fused INT4 GEMM kernel — 17x decode speedup
Replace the INT4 dequant→sanitize→bf16_gemm_tiled pipeline (4 kernels,
52MB allocation, 52MB intermediate buffer per GEMM) with the existing
fused int4_gemm_auto_round kernel that reads qweight directly and
dequants in registers.

INT4 decode: 2.2s/token → 0.130s/token (17x improvement)
INT4 prefill: 2.5s → 0.288s (8.6x improvement)

Correctness verified: oracle comparison 0 FAIL, 38 PASS, 27 WARN,
final logits cosine 0.99478 (unchanged from dequant→GEMM path).

NVFP4 still uses dequant→GEMM path (2.2s/token) — fused kernel pending.

Also fixed: scales parameter type mismatch in launch_int4_gemm_auto_round
(CudaSlice<f16> instead of CudaSlice<bf16>).

+++ crates/cuda/src/oxide_bridge.rs
@@ -880,7 +880,7 @@ impl OxideKernels {
         stream: &Arc<CudaStream>,
         output: &mut CudaSlice<half::bf16>,
         weight: &CudaSlice<u32>,
-        scales: &CudaSlice<half::bf16>,
+        scales: &CudaSlice<half::f16>,
         zeros: &CudaSlice<u32>,
         input: &CudaSlice<half::bf16>,
         m: u32,
